Add external links and matching certificates from either Route or Ingress if able to do so.

+++ tests/storage/export.go
@@ -21,6 +21,7 @@ package storage
 
 import (
 	"context"
+	"crypto/x509"
 	"fmt"
 	"path/filepath"
 	"strings"
@@ -30,6 +31,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	k8sv1 "k8s.io/api/core/v1"
+	networkingv1 "k8s.io/api/networking/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -41,12 +43,16 @@ import (
 	"kubevirt.io/client-go/kubecli"
 	"kubevirt.io/client-go/log"
 	cdiv1 "kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1"
-
+	"kubevirt.io/kubevirt/pkg/certificates/triple/cert"
+	"kubevirt.io/kubevirt/pkg/virt-operator/resource/generate/components"
 	"kubevirt.io/kubevirt/tests"
 	cd "kubevirt.io/kubevirt/tests/containerdisk"
 	"kubevirt.io/kubevirt/tests/flags"
+	"kubevirt.io/kubevirt/tests/framework/checks"
 	"kubevirt.io/kubevirt/tests/libstorage"
 	"kubevirt.io/kubevirt/tests/util"
+
+	routev1 "github.com/openshift/api/route/v1"
 )
 
 const (
@@ -67,6 +73,32 @@ const (
 	pvcNotFoundReason = "pvcNotFound"
 
 	proxyUrlBase = "https://virt-exportproxy.%s.svc/api/export.kubevirt.io/v1alpha1/namespaces/%s/virtualmachineexports/%s%s"
+
+	tlsKey   = "tls.key"
+	tlsCert  = "tls.crt"
+	testKey  = "test"
+	testCert = `-----BEGIN CERTIFICATE-----
+MIIDJTCCAg2gAwIBAgIUYhSmCxHywX8qmhCPSZweno9MP8cwDQYJKoZIhvcNAQEL
+BQAwIjEgMB4GA1UEAwwXdm1leHBvcnQtcHJveHkudGVzdC5uZXQwHhcNMjIwNjEz
+MTcwNTA4WhcNMjMwNjEzMTcwNTA4WjAiMSAwHgYDVQQDDBd2bWV4cG9ydC1wcm94
+eS50ZXN0Lm5ldDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALrkCWbD
+z2i/QAOQXOX6bk0Q3wq6wRz+qHtr9mu93FQrZUcMzYpr/dLMaW18PVes4RjMuqjt
+cbTjgGIwoS3VFCyub9wB1cro+0avf8cnl7Z10RPA7i6b2UV800HcwlmrbGr29kN2
+E2x2PajAAjoAqAaqu1rhP1ua7xynpWiDk64pXVDxdPUUvyoDD6eL3KqSqA1X/iys
+fG6s+ypL540mIDLR5rxtvvdD4J7hktT+UIinbEE7Q9DwoGJ6xKqYagOb2CLymNLZ
+ruJ3vmiTqd/2HXGI+4xe2fR6v23CRhLIeRBIzznIJOwBGe/YutGCIHa35AATlhwS
+HzFagYwOATIMvt8CAwEAAaNTMFEwHQYDVR0OBBYEFJVH/4uLVQ8gPqa6x7D46Jga
+jikIMB8GA1UdIwQYMBaAFJVH/4uLVQ8gPqa6x7D46JgajikIMA8GA1UdEwEB/wQF
+MAMBAf8wDQYJKoZIhvcNAQELBQADggEBAKPg9lwsRMpJ6wR1hR/1GWFVqJsb7t7G
+nr0SPsYBYzOkIfMtIK7kRfLQgMRTNOpvyC13zkDeAFjejr+Ovl7d4ONLGP/7qcj/
+dLN0Tas2m0pMJjxJa40WYU+XlnO47R/HtYnrn7RMIZ2TUroisf2oiEUJ36989mp5
+ArFLLJzozIrpAP5lHThpJjQCs6wM/d7daasRBfXucGPMvXcUnVmGo+2pUjTourpG
+NH18UMKg4B0B1Gs6LsQVYrQerP9zI14p3nu3ieBFEUT2wuERycbHLwCH+mC4HpbD
+pS/x4m6CgALv0lrhHpffeOc8lncNEzkZbRpCG1NBbPA6mFLtklcuER8=
+-----END CERTIFICATE-----`
+	// This hostname matches the cert above (testCert)
+	testHostName     = "vmexport-proxy.test.net"
+	subjectAltNameId = "2.5.29.17"
 )
 
 var _ = SIGDescribe("Export", func() {
@@ -93,8 +125,10 @@ var _ = SIGDescribe("Export", func() {
 	})
 
 	AfterEach(func() {
-		err := virtClient.CoreV1().Secrets(token.Namespace).Delete(context.Background(), token.Name, metav1.DeleteOptions{})
-		Expect(err).ToNot(HaveOccurred())
+		if token != nil {
+			err := virtClient.CoreV1().Secrets(token.Namespace).Delete(context.Background(), token.Name, metav1.DeleteOptions{})
+			Expect(err).ToNot(HaveOccurred())
+		}
 	})
 
 	addBlockVolume := func(pod *k8sv1.Pod, volumeName string) *k8sv1.Pod {
@@ -605,6 +639,21 @@ var _ = SIGDescribe("Export", func() {
 		return export
 	}
 
+	matchesCNOrAlt := func(cert *x509.Certificate, hostName string) bool {
+		fmt.Fprintf(GinkgoWriter, "CN: %s, hostname: %s\n", cert.Subject.CommonName, hostName)
+		if strings.Contains(cert.Subject.CommonName, hostName) {
+			return true
+		}
+		for _, extension := range cert.Extensions {
+			fmt.Fprintf(GinkgoWriter, "ExtensionID: %s, subjectAltNameId: %s, value: %s, hostname: %s\n", extension.Id.String(), subjectAltNameId, string(extension.Value), hostName)
+			if extension.Id.String() == subjectAltNameId && strings.Contains(string(extension.Value), hostName) {
+				return true
+			}
+		}
+
+		return false
+	}
+
 	It("Should recreate the exporter pod and secret if the pod fails", func() {
 		sc, exists := libstorage.GetRWOFileSystemStorageClass()
 		if !exists {
@@ -773,5 +822,143 @@ var _ = SIGDescribe("Export", func() {
 			Expect(err).ToNot(HaveOccurred())
 			return p.Status.Phase == k8sv1.PodSucceeded
 		}, 90*time.Second, 1*time.Second).Should(BeTrue())
+	})
+
+	Context("Ingress", func() {
+		const (
+			tlsSecretName = "test-tls"
+		)
+
+		AfterEach(func() {
+			err := virtClient.CoreV1().Secrets(flags.KubeVirtInstallNamespace).Delete(context.Background(), tlsSecretName, metav1.DeleteOptions{})
+			if !errors.IsNotFound(err) {
+				Expect(err).ToNot(HaveOccurred())
+			}
+			err = virtClient.NetworkingV1().Ingresses(flags.KubeVirtInstallNamespace).Delete(context.Background(), "export-proxy-ingress", metav1.DeleteOptions{})
+			if !errors.IsNotFound(err) {
+				Expect(err).ToNot(HaveOccurred())
+			}
+		})
+
+		createIngressTLSSecret := func(name string) {
+			secret := &k8sv1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      name,
+					Namespace: flags.KubeVirtInstallNamespace,
+				},
+				StringData: map[string]string{
+					tlsKey:  testKey,
+					tlsCert: testCert,
+				},
+			}
+			_, err := virtClient.CoreV1().Secrets(flags.KubeVirtInstallNamespace).Create(context.Background(), secret, metav1.CreateOptions{})
+			Expect(err).ToNot(HaveOccurred())
+		}
+
+		createIngress := func(tlsSecretName string) *networkingv1.Ingress {
+			prefix := networkingv1.PathTypePrefix
+			ingress := &networkingv1.Ingress{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "export-proxy-ingress",
+					Namespace: flags.KubeVirtInstallNamespace,
+				},
+				Spec: networkingv1.IngressSpec{
+					IngressClassName: pointer.StringPtr("ingress-class-name"),
+					DefaultBackend: &networkingv1.IngressBackend{
+						Service: &networkingv1.IngressServiceBackend{
+							Name: "virt-exportproxy",
+							Port: networkingv1.ServiceBackendPort{
+								Number: int32(443),
+							},
+						},
+					},
+					TLS: []networkingv1.IngressTLS{
+						{
+							Hosts: []string{
+								testHostName,
+							},
+							SecretName: tlsSecretName,
+						},
+					},
+					Rules: []networkingv1.IngressRule{
+						{
+							Host: testHostName,
+							IngressRuleValue: networkingv1.IngressRuleValue{
+								HTTP: &networkingv1.HTTPIngressRuleValue{
+									Paths: []networkingv1.HTTPIngressPath{
+										{
+											Path:     "/",
+											PathType: &prefix,
+											Backend: networkingv1.IngressBackend{
+												Service: &networkingv1.IngressServiceBackend{
+													Name: "virt-exportproxy",
+													Port: networkingv1.ServiceBackendPort{
+														Number: int32(443),
+													},
+												},
+											},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			}
+			ingress, err := virtClient.NetworkingV1().Ingresses(flags.KubeVirtInstallNamespace).Create(context.Background(), ingress, metav1.CreateOptions{})
+			Expect(err).ToNot(HaveOccurred())
+			return ingress
+		}
+
+		It("should populate external links and cert and contain ingress host", func() {
+			sc, exists := libstorage.GetRWOFileSystemStorageClass()
+			if !exists {
+				Skip("Skip test when Filesystem storage is not present")
+			}
+			createIngressTLSSecret(tlsSecretName)
+			ingress := createIngress(tlsSecretName)
+			vmExport := createRunningExport(sc, k8sv1.PersistentVolumeFilesystem)
+			Expect(vmExport.Status.Links.External.Cert).To(Equal(testCert))
+			certs, err := cert.ParseCertsPEM([]byte(vmExport.Status.Links.External.Cert))
+			Expect(err).ToNot(HaveOccurred())
+			Expect(len(certs)).To(Equal(1))
+			prefix := fmt.Sprintf("%s-%s", components.VirtExportProxyServiceName, flags.KubeVirtInstallNamespace)
+			domainName := strings.TrimPrefix(ingress.Spec.Rules[0].Host, prefix)
+			Expect(matchesCNOrAlt(certs[0], domainName)).To(BeTrue())
+			Expect(vmExport.Status.Links.External.Volumes[0].Formats[0].Url).To(ContainSubstring(ingress.Spec.Rules[0].Host))
+		})
+	})
+
+	Context("Route", func() {
+		getExportRoute := func() *routev1.Route {
+			route, err := virtClient.RouteClient().Routes(flags.KubeVirtInstallNamespace).Get(context.Background(), components.VirtExportProxyServiceName, metav1.GetOptions{})
+			Expect(err).ToNot(HaveOccurred())
+			return route
+		}
+
+		It("should populate external links and cert and contain route host", func() {
+			sc, exists := libstorage.GetRWOFileSystemStorageClass()
+			if !exists {
+				Skip("Skip test when Filesystem storage is not present")
+			}
+			if !checks.IsOpenShift() {
+				Skip("Not on openshift")
+			}
+			vmExport := createRunningExport(sc, k8sv1.PersistentVolumeFilesystem)
+			certs, err := cert.ParseCertsPEM([]byte(vmExport.Status.Links.External.Cert))
+			Expect(err).ToNot(HaveOccurred())
+			Expect(len(certs)).To(Equal(1))
+			route := getExportRoute()
+			host := ""
+			if len(route.Status.Ingress) > 0 {
+				host = route.Status.Ingress[0].Host
+			}
+			Expect(host).ToNot(BeEmpty())
+			prefix := fmt.Sprintf("%s-%s", components.VirtExportProxyServiceName, flags.KubeVirtInstallNamespace)
+			domainName := strings.TrimPrefix(host, prefix)
+			Expect(matchesCNOrAlt(certs[0], domainName)).To(BeTrue())
+			Expect(vmExport.Status.Links.External.Volumes[0].Formats[0].Url).To(ContainSubstring(host))
+
+		})
 	})
 })
